rework POT check and passing of the block processing width for special case of inclusive scan

+++ Shared/MetalPrefixSumRenderFrame.h
@@ -24,7 +24,7 @@
 @property (nonatomic, assign) int width;
 @property (nonatomic, assign) int height;
 
-@property (nonatomic, assign) int blockDim;
+@property (nonatomic, assign) NSUInteger blockDim;
 
 // The original input image order in image order
